MOved satfunc initializers to separate file.
