+++ opm/parser/eclipse/CMakeLists.txt
@@ -179,6 +179,7 @@ EclipseState/Grid/EclipseGrid.hpp
 EclipseState/Grid/GridProperty.hpp
 EclipseState/Grid/GridProperties.hpp
 EclipseState/Grid/GridPropertyInitializers.hpp
+EclipseState/Grid/SatfuncPropertyInitializers.hpp
 EclipseState/Grid/Box.hpp
 EclipseState/Grid/BoxManager.hpp
 EclipseState/Grid/FaceDir.hpp
