+++ opm/parser/eclipse/EclipseState/EclipseState.cpp
@@ -30,6 +30,8 @@
 #include <opm/parser/eclipse/EclipseState/Grid/GridProperties.hpp>
 #include <opm/parser/eclipse/EclipseState/Grid/Box.hpp>
 #include <opm/parser/eclipse/EclipseState/Grid/BoxManager.hpp>
+#include <opm/parser/eclipse/EclipseState/Grid/SatfuncPropertyInitializers.hpp>
+
 #include <opm/parser/eclipse/OpmLog/OpmLog.hpp>
